Add information surrounding day 19
After many frustrating hours, I went ahead and just read someone else's answer and implemented it as I understood it.  I sort of see this as too close to the line of cheating, so I'm not committing this day's answer; instead, feel free to browse two of my many failed attempts to even get started.